migrate to markitdown for proccessing docs

+++ backend/src/service/extraction.py
@@ -1,16 +1,17 @@
 import os
 from typing import List
 from src.lib.logger.index import logger
-import pymupdf4llm
-import pymupdf
 from src.lib.youtube.index import client as youtube_client
 from src.lib.firecrawl.index import client as firecrawl_client
 from src.lib.openai.index import client as openai_client
+from markitdown import MarkItDown
+from pypdf import PdfReader
+
+md_client = MarkItDown()
 
 
 class ExtractionService:
     def __init__(self):
-        pymupdf.pro.unlock()
         logger.info("EXTRACTION SERVICE INITIALIZED!")
 
     def extract_youtube_transcript(self, video_id: str):
@@ -40,9 +41,11 @@ class ExtractionService:
         logger.info(f"Website content extracted successfully: markdown is {markdown}")
         return markdown, metadata
 
-    def extract_document_content(self, file_path: str) -> List[dict]:
+    def extract_document_content(
+        self, file_path: str
+    ) -> List[dict]:  # returns a list of {"text": "", "pageNumber": int}
         """
-        Extracts the content of a document from the provided file path. Documents are PDFs, DOCXs, and PPTXs.
+        Extracts the content of a document from the provided file path. Documents are any office documents supported by markitdown.
 
         Args:
             file_path (str): The file path of the document to extract content from.
@@ -50,11 +53,24 @@ class ExtractionService:
         Returns:
             list: The extracted content of the document.
         """
+        if not file_path or not os.path.exists(file_path):
+            raise ValueError("File path is required to process document")
         try:
             # extract Markdown for each page as a list of dictionaries
-            data = pymupdf4llm.to_markdown(
-                file_path, page_chunks=True
-            )  # returns a list of {"text": "", "pageNumber": int}
+            isPdf = file_path.endswith(".pdf")
+            if isPdf:
+                reader = PdfReader(file_path)
+                pages = reader.pages
+                data = []
+                for page in pages:
+                    data.append(
+                        {"text": page.extract_text(), "pageNumber": page.page_number}
+                    )
+            else:
+                data = md_client.convert(file_path)
+                logger.info(f"Document content extracted successfully: data is {data}")
+                data = [{"text": data.text_content, "pageNumber": 1}]
+                logger.info(f"Document content extracted successfully: data is {data}")
 
             os.remove(file_path)
             if not data:
